Initial work on Indium; basic triangle rendering!
Next up: initial work on the shader translator (AIR-to-SPIR-V)

Ok, so this code is almost certainly not very efficient or even fully
compliant, but it works! We've got a basic triangle render
roughly translated from Apple's "Using a Render Pipeline to Render
Primitives" sample working with Indium.

Rendering in a loop (what you're supposed to do) is currently broken due
to the fact that we don't have a proper way to do vsync yet. It also
produces some Vulkan validation errors due to (apparently) destroying
buffers/semaphores before Vulkan is done with them.