Some more work on Iridium; basic outline generation working
The hardest part is done: getting SPIR-V code generated and setting up
the module interface. From here, it should be straightforward to just
translate the remaining instructions appropriately and everything should work.

Note that the current code is only designed for a specific case (the
shader from Apple's "Using a Render Pipeline to Render Primitives" example).
We'll focus on this case for now and expand to support other inputs later.