Untested code to generate a module for the example shader
This code is completely untested, except for the fact that it generates
valid SPIR-V that passes the checks performed by `spirv-val`.

This is just a sample implementation, just to prove to myself that this
is in fact possible. I'll rework this later to use physical storage
buffer addressing, since Metal code expects to be able to play around
with pointers. Plus, it's honestly just easier to use Vulkan's BDA
(buffer device address) support; it'll make translating the LLVM IR
much easier.