A cube renders!
With this new example, I've found a couple more bugs in the current
implementation, which is good that I found them now, the problem is that
I have no clue why they happen. The first is a random segfault in the
CommandBuffer::commit method which occurred because `_commandBuffer` was
`nullptr` for some reason. The second is a bug where the completion handlers
of command buffers will sometimes not be invoked; this seems only to occur
when the render target is hidden.