iridium: Add support for some more LLVM IR
One major hurdle is going to be detecting CFG structures and marking them.
In its current form, translated shaders with branching in them fail to
validate. This is because SPIR-V requires structured CFG declarations. The
problem for us is that LLVM IR includes no such information.

The unfortunate and error-prone approach that we'll have to take is we'll
have to apply some heuristics to detect and mark CFG structures such as
ifs/selections, switches, loops, etc.