iridium: Ensure bitcasted pointers use the same storage class
While spirv-val doesn't complain, it's not correct to bitcast between
storage classes and then load from the bitcasted pointer (which has the
wrong storage class).

In the future, once I decipher what each address space in the LLVM IR is
(because the address spaces are AIR-specific), we can probably just use
that to infer the correct storage class. Until then, we'll have to keep
hacking it a little bit like this.