Dynamically link to Vulkan at runtime
This allows us to (partially) gracefully handle the case when a user
doesn't have a Vulkan runtime available. The other missing part of this
is to gracefully when we fail to create an instance.

Vulkan headers are still required at build-time but no library is
strictly required (neither at build-time nor at runtime).