Demote some required extensions to optional device features
Instead of causing device creation to fail, not having one of these
extensions will simply mean that the associated features are unavailable.
For example, semaphores can only be created as exportable if the device
supports external semaphore FDs.

This allows Indium to work with certain devices such as compute-only
devices or software renderers like lavapipe/llvmpipe.